+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/.settings/org.eclipse.jpt.core.prefs
