+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/.settings/org.eclipse.wst.common.component
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="dackelCMDB">
+    <wb-module deploy-name="DackelCMDBJBoss">
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-        <property name="context-root" value="dackelCMDB"/>
         <property name="java-output-path" value="/dackelCMDB/target/classes"/>
+        <property name="context-root" value="DackelCMDBJBoss"/>
     </wb-module>
 </project-modules>
