+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <faceted-project>
   <fixed facet="wst.jsdt.web"/>
-  <installed facet="java" version="1.7"/>
   <installed facet="jst.web" version="2.5"/>
   <installed facet="jpt.jpa" version="2.1"/>
   <installed facet="jst.cdi" version="1.2"/>
@@ -9,4 +8,5 @@
   <installed facet="jst.jsf" version="2.2"/>
   <installed facet="jst.jaxrs" version="2.0"/>
   <installed facet="wst.jsdt.web" version="1.0"/>
+  <installed facet="java" version="1.8"/>
 </faceted-project>
