+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/.settings/org.hibernate.eclipse.console.prefs
