+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/Model.notation
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<notation:Diagram xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:notation="http://www.eclipse.org/gmf/runtime/1.0.2/notation" xmlns:style="http://www.eclipse.org/papyrus/infra/gmfdiag/style" xmlns:uml="http://www.eclipse.org/uml2/5.0.0/UML" xmi:id="_G_N5UE8GEemZZc8tmMgLxA" type="PapyrusUMLClassDiagram" name="UML Class Diagramm" measurementUnit="Pixel">
+  <styles xmi:type="notation:StringValueStyle" xmi:id="_G_N5UU8GEemZZc8tmMgLxA" name="diagram_compatibility_version" stringValue="1.4.0"/>
+  <styles xmi:type="notation:DiagramStyle" xmi:id="_G_N5Uk8GEemZZc8tmMgLxA"/>
+  <styles xmi:type="style:PapyrusDiagramStyle" xmi:id="_G_N5U08GEemZZc8tmMgLxA" diagramKindId="org.eclipse.papyrus.uml.diagram.class">
+    <owner xmi:type="uml:Model" href="Model.uml#_oqeYkEzNEemXDdDFkUPe3A"/>
+  </styles>
+  <element xmi:type="uml:Model" href="Model.uml#_oqeYkEzNEemXDdDFkUPe3A"/>
+</notation:Diagram>
