+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/Model_de_DE.properties
@@ -1,0 +1,2 @@
+#Wed Mar 27 21:42:17 CET 2019
+_label_dackelcmdb=DackelCMDB
