+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/pom.xml
@@ -19,14 +19,14 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>de.percsi.products</groupId>
-    <artifactId>dackelCMDB</artifactId>
+    <groupId>de.percsi.products.dackelcmdb.jboss</groupId>
+    <artifactId>DackelCMDB</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>war</packaging>
-    <name>WildFly Quickstarts: dackelCMDB</name>
-    <description>A starter Java EE 7 webapp project for use on JBoss WildFly / WildFly, generated from the jboss-javaee6-webapp archetype</description>
+    <name>DackelCMDB Projekt on JBoss</name>
+    <description>DackelCMDB - Configuration Management Datenbank Projekt</description>
 
-    <url>http://wildfly.org</url>
+    <url>istvanpercsi.github.io/DackelCMDB</url>
     <licenses>
         <license>
             <name>Apache License, Version 2.0</name>
