+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/src/main/webapp/WEB-INF/beans.xml
