+++ projects/DackelCMDB-JavaBackend/DackelCMDBJBoss/src/main/webapp/index.html
